# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tenants": {
      "post": {
        "description": "Creates a tenant (an application) and mints its API key. The key is returned exactly once and cannot be retrieved afterwards — only a bcrypt hash is stored.\n\nGuarded by the `x-admin-key` header whenever `ADMIN_API_KEY` is configured. It is unset locally and in CI, so the guard is a no-op there; Terraform sets it from Secret Manager in staging and production, where the service is publicly reachable and an open registration endpoint would let anyone provision themselves credentials.",
        "operationId": "TenantsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tenant created. Store `apiKey` now — it is not shown again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTenantResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (e.g. slug is not lowercase-alphanumeric).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `x-admin-key` (when admin auth is enabled).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "A tenant with this name or slug already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin-key": []
          }
        ],
        "summary": "Register a new tenant",
        "tags": [
          "Tenants"
        ]
      }
    },
    "/api/v1/tenants/me": {
      "get": {
        "description": "There is deliberately no endpoint that lists all tenants: it is not required, and on a public URL it would enumerate every tenant on the platform. API key material (`apiKeyHash`, `apiKeyLookup`) is marked `select: false` on the entity and never appears in any response.",
        "operationId": "TenantsController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The authenticated tenant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `x-api-key`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Get the tenant owning the presented API key",
        "tags": [
          "Tenants"
        ]
      }
    },
...
  },
  "info": {
    "title": "Multi-Tenant Feature Flag Service",
...
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local (Docker Compose)"
    },
    {
      "url": "https://feature-flag-service-staging-dhqwxiey3q-uc.a.run.app",
      "description": "Staging (Cloud Run)"
    },
    {
      "url": "https://feature-flag-service-production-<hash>-uc.a.run.app",
      "description": "Production (Cloud Run)"
    }
  ],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Tenant API key, returned once at registration. Identifies the tenant for every scoped operation."
      },
      "admin-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key",
        "description": "Platform admin key, required for tenant registration wherever ADMIN_API_KEY is configured. Read it with: gcloud secrets versions access latest --secret=feature-flag-service-<env>-admin-api-key"
      }
    },
    "schemas": {
      "CreateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Human-readable application name. Must be unique across the platform.",
            "example": "Acme Corporation"
          },
          "slug": {
            "type": "string",
            "description": "URL-safe identifier. Lowercase alphanumeric and hyphens only. Must be unique.",
            "example": "acme-corp",
            "pattern": "^[a-z0-9-]+$"
          },
          "metadata": {
            "type": "object",
            "description": "Arbitrary key/value metadata stored against the tenant. Not interpreted by the service.",
            "example": {
              "team": "payments",
              "costCentre": "CC-1042"
            },
            "additionalProperties": true
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "CreateTenantResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Acme Corporation"
          },
          "slug": {
            "type": "string",
            "example": "acme-corp"
          },
          "apiKey": {
            "type": "string",
            "description": "The plaintext API key. Returned ONCE, at creation, and never retrievable again — only a bcrypt hash is stored.",
            "example": "ffs_EXAMPLE0000000000000000000000"
          },
          "message": {
            "type": "string",
            "example": "Store this API key securely — it will not be shown again"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "apiKey",
          "message",
          "createdAt"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "description": "Either a single message or, for validation failures, one entry per invalid field.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": [
              "environment must be one of the following values: development, staging, production"
            ]
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
      "TenantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Acme Corporation"
          },
          "slug": {
            "type": "string",
            "example": "acme-corp"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "example": {}
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "isActive",
          "metadata",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}